pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "openai>=1.97.0",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
]
